+++ pages/desktop/developers/edit_addon.py
