+++ pages/desktop/frontend/static_pages.py
